requirements.txt
paho-mqtt
pyusb
